Uploaded new PAS and updated version for v1.4 release.

+++ pascal/Megapacker.pas
@@ -200,7 +200,7 @@ end;
 
 procedure TCODEC.CompressedStreamWriteBits(Value: string);
 var
-  I: Byte;
+  I: Integer;
 begin
   for I := 1 to Length(Value) do
   begin
